requirements.txt
fastapi==0.110.0
pydantic==2.6.4
pyjwt==2.8